Product model was created and it is already connected to the database
